Split API handlers into domain specific files
